dev-requirements.txt
-r requirements.txt
-r docs/requirements.txt
coverage
pytest
pyyaml
virtualenv
